🐛 fix: eliminate runtime warnings surfaced during smoke tests
- Eames2012.py make_reciprocal: add np.asarray(census, dtype=np.float64)
  before positional indexing to silence pandas FutureWarning (Series
  integer access will always use label-based lookup in pandas 3.x).
- seir_environment.py: cast budget element to np.float32 before
  concatenation so highs_per_agent stays float32 end-to-end; cast
  observation_space_per_agent bounds explicitly to float32 to suppress
  gymnasium Box precision-downcast warnings.
- wrappers.py NormalizedObservationWrapper: create low/high bounds with
  dtype=np.float32 to avoid the same Box downcast warning.
- seir_environment_single_sb_ppo.py / seir_environment_multi_sb_ppo.py:
  pass device="cpu" to PPO — MlpPolicy trains faster on CPU than GPU and
  this silences the stable-baselines3 device warning.

+++ epcontrol/wrappers.py
@@ -206,8 +206,9 @@ class NormalizedObservationWrapper(ObservationWrapper):
             raise AssertionError(
                 "This wrapper cannot be used for observation spaces with an infinite lower/upper bound.")
         self.observation_space: Box = Box(
-            low=np.zeros(self.env.observation_space.shape),
-            high=np.ones(self.env.observation_space.shape)
+            low=np.zeros(self.env.observation_space.shape, dtype=np.float32),
+            high=np.ones(self.env.observation_space.shape, dtype=np.float32),
+            dtype=np.float32,
         )
 
     def observation(self, observation: np.ndarray) -> np.ndarray:
